requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
requests
assemblyai
google-generativeai
python-multipart